moved storage to util in internal, ref binding
